+++ packages/solutions/app-tools/tests/builder/createToolsConfig.test.ts
@@ -1,0 +1,33 @@
+import { createBuilderTsChecker } from '../../src/builder/createToolsConfig';
+
+describe('test create Builder TsChecker Config', () => {
+  it('should disableTsChecker', () => {
+    const output1 = {
+      enableTsLoader: true,
+    };
+    const tsCheckerConfig1 = createBuilderTsChecker(output1);
+    expect(tsCheckerConfig1).toBeFalsy();
+
+    const output2 = {
+      disableTsChecker: true,
+    };
+    const tsCheckerConfig2 = createBuilderTsChecker(output2);
+    expect(tsCheckerConfig2).toBeFalsy();
+  });
+
+  it('should return default TsChecker Config', () => {
+    const output = {};
+    const tsCheckerConfig = createBuilderTsChecker(output);
+    expect(tsCheckerConfig).toEqual({
+      issue: {
+        include: [{ file: '**/src/**/*' }],
+        exclude: [
+          { file: '**/*.(spec|test).ts' },
+          { file: '**/node_modules/**/*' },
+        ],
+      },
+    });
+  });
+});
+
+// TODO: other test