+++ tests/integration/css/fixtures/css-ts-declaration/src/App.jsx
@@ -1,8 +1,8 @@
-import React from 'react';
-import { redText } from './index.module.css';
+// import React from 'react';
+import style from './index.module.css';
 
 const App = () => (
-  <div id="verify-red" className={redText}>
+  <div id="verify-red" className={style.redText}>
     This text should be red.
   </div>
 );