+++ tests/integration/runtime/fixtures/use-loader/modern.config.js
@@ -1,7 +1,4 @@
 module.exports = {
-  runtime: {
-    router: true,
-  },
   server: {
     ssr: true,
   },