+++ packages/generator/new-action/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @modern-js/new-action
 
+## 1.21.1
+
+### Patch Changes
+
+- @modern-js/generator-common@2.4.6
+- @modern-js/generator-utils@2.4.6
+- @modern-js/utils@1.21.1
+
 ## 1.21.0
 
 ### Patch Changes